+++ src/game/systems/behavior-system.cpp
@@ -26,7 +26,7 @@ behavior_system::behavior_system(entity::registry& registry):
 	updatable_system(registry)
 {}
 
-void behavior_system::update(double t, double dt)
+void behavior_system::update(float t, float dt)
 {
 	// registry.view<behavior_component>().each(
 		// [&](entity::id entity_id, auto& behavior)
