+++ src/game/systems/metamorphosis-system.hpp
@@ -28,7 +28,7 @@ class metamorphosis_system:
 {
 public:
 	explicit metamorphosis_system(entity::registry& registry);
-	virtual void update(double t, double dt);
+	virtual void update(float t, float dt);
 	
 	/**
 	 * Sets the factor by which the timestep `dt` will be scaled.
